requirements.txt
streamlit
spacy
sentence-transformers
faiss-cpu
PyMuPDF
numpy